+++ src/components/MeWebsite.vue
@@ -2,11 +2,14 @@
 // 跳转到官网: Redis中英文/Valkey中英文
 import { openUrl } from '@tauri-apps/plugin-opener'
 
-import { isZh, meOk } from '@/utils/util'
+import { isZh } from '@/utils/util'
+
+/** 与 `<me-website to="…">`、DOC_PATHS 键一致（供外部 props） */
+type DocTopic = 'info' | 'config' | 'client' | 'command' | 'slowlog' | 'monitor' | 'pubsub'
 
 const props = withDefaults(
   defineProps<{
-    to: string
+    to: DocTopic
     placement?: string
     marginLeft?: string
   }>(),
@@ -16,76 +19,59 @@ const props = withDefaults(
   },
 )
 
-// 官网站点
-const websize = {
+/** 下拉项 command，与模板中 el-dropdown-item 一致 */
+const WEB_ORIGIN = {
   redis: 'https://redis.io',
-  'redis-zh': 'https://redis.ac.cn',
+  redisZh: 'https://redis.ac.cn',
   valkey: 'https://valkey.io',
-  'valkey-zh': 'https://valkey.cn',
-}
+  valkeyZh: 'https://valkey.cn',
+} as const
 
-// 信息命令
-const info = {
-  redis: '/docs/latest/commands/info/',
-  valkey: '/commands/info/',
-}
+type SiteCmd = keyof typeof WEB_ORIGIN
 
-// 配置文件
-const config = {
-  redis: '/docs/latest/operate/oss_and_stack/management/config/',
-  valkey: '/topics/valkey.conf/',
-}
+/** `to` → redis / valkey 路径后缀 */
+const DOC_PATHS = {
+  info: {
+    redis: '/docs/latest/commands/info/',
+    valkey: '/commands/info/',
+  },
+  config: {
+    redis: '/docs/latest/operate/oss_and_stack/management/config/',
+    valkey: '/topics/valkey.conf/',
+  },
+  client: {
+    redis: '/docs/latest/commands/client-list/',
+    valkey: '/commands/client-list/',
+  },
+  command: {
+    redis: '/docs/latest/commands/',
+    valkey: '/commands/',
+  },
+  slowlog: {
+    redis: '/docs/latest/commands/slowlog-get/',
+    valkey: '/commands/slowlog-get/',
+  },
+  monitor: {
+    redis: '/docs/latest/commands/monitor/',
+    valkey: '/commands/monitor/',
+  },
+  pubsub: {
+    redis: '/docs/latest/commands/psubscribe/',
+    valkey: '/commands/psubscribe/',
+  },
+} as const satisfies Record<DocTopic, { readonly redis: string; readonly valkey: string }>
 
-// 客户端列表
-const client = {
-  redis: '/docs/latest/commands/client-list/',
-  valkey: '/commands/client-list/',
-}
-
-// 命令列表
-const command = {
-  redis: '/docs/latest/commands/',
-  valkey: '/commands/',
-}
-
-// 慢日志命令
-const slowlog = {
-  redis: '/docs/latest/commands/slowlog-get/',
-  valkey: '/commands/slowlog-get/',
-}
-
-// 监控命令
-const monitor = {
-  redis: '/docs/latest/commands/monitor/',
-  valkey: '/commands/monitor/',
-}
-
-// 发布订阅命令
-const pubsub = {
-  redis: '/docs/latest/commands/psubscribe/',
-  valkey: '/commands/psubscribe/',
-}
+type Vendor = keyof (typeof DOC_PATHS)['info']
 
 function handleCommand(cmd: string): void {
-  const part = cmd.split('-')[0] as keyof typeof info
-  const base = websize[cmd as keyof typeof websize]
-  void openUrl(base + info[part])
-  if (props.to === 'info') {
-    void openUrl(base + info[part])
-  } else if (props.to === 'config') {
-    void openUrl(base + config[part])
-  } else if (props.to === 'client') {
-    void openUrl(base + client[part])
-  } else if (props.to === 'command') {
-    void openUrl(base + command[part])
-  } else if (props.to === 'slowlog') {
-    void openUrl(base + slowlog[part])
-  } else if (props.to === 'monitor') {
-    void openUrl(base + monitor[part])
-  } else if (props.to === 'pubsub') {
-    void openUrl(base + pubsub[part])
-  } else {
-    meOk(`TODO: ${cmd}`)
+  const site = cmd as SiteCmd
+  const vendor = (site.endsWith('Zh') ? site.slice(0, -2) : site) as Vendor
+  const base = WEB_ORIGIN[site]
+  const paths = DOC_PATHS[props.to]
+
+  void openUrl(base + DOC_PATHS.info[vendor])
+  if (props.to !== 'info') {
+    void openUrl(base + paths[vendor])
   }
 }
 </script>
@@ -96,17 +82,17 @@ function handleCommand(cmd: string): void {
     <template #dropdown>
       <el-dropdown-menu>
         <el-dropdown-item command="redis">
-          <me-icon icon="me-icon-redis" name="Redis"
-        /></el-dropdown-item>
-        <el-dropdown-item command="valkey"
-          ><me-icon icon="me-icon-valkey" name="Valkey"
-        /></el-dropdown-item>
-        <el-dropdown-item command="redis-zh" v-if="isZh"
-          ><me-icon icon="me-icon-redis" name="Redis 中文"
-        /></el-dropdown-item>
-        <el-dropdown-item command="valkey-zh" v-if="isZh"
-          ><me-icon icon="me-icon-valkey" name="Valkey 中文"
-        /></el-dropdown-item>
+          <me-icon icon="me-icon-redis" name="Redis" />
+        </el-dropdown-item>
+        <el-dropdown-item v-if="isZh" command="redisZh">
+          <me-icon icon="me-icon-redis" name="Redis 中文" />
+        </el-dropdown-item>
+        <el-dropdown-item command="valkey">
+          <me-icon icon="me-icon-valkey" name="Valkey" />
+        </el-dropdown-item>
+        <el-dropdown-item v-if="isZh" command="valkeyZh">
+          <me-icon icon="me-icon-valkey" name="Valkey 中文" />
+        </el-dropdown-item>
       </el-dropdown-menu>
     </template>
   </el-dropdown>
